+++ src/test/java/dev/marksman/enhancediterables/NonEmptyFiniteIterableTest.java
@@ -18,6 +18,7 @@ import static com.jnape.palatable.lambda.adt.hlist.HList.tuple;
 import static com.jnape.palatable.lambda.functions.builtin.fn1.Constantly.constantly;
 import static com.jnape.palatable.lambda.functions.builtin.fn1.Id.id;
 import static com.jnape.palatable.lambda.functions.builtin.fn2.Tupler2.tupler;
+import static com.jnape.palatable.lambda.functor.builtin.Lazy.lazy;
 import static dev.marksman.enhancediterables.NonEmptyFiniteIterable.nonEmptyFiniteIterable;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
@@ -183,6 +184,18 @@ class NonEmptyFiniteIterableTest {
     }
 
     @Nested
+    @DisplayName("distinct")
+    class Distinct {
+
+        @Test
+        void removesRepeatedElementsAndRetainsOrder() {
+            assertThat(nonEmptyFiniteIterable(1, asList(2, 2, 3, 3, 3, 2, 2, 1, 4)).distinct(),
+                    contains(1, 2, 3, 4));
+        }
+
+    }
+
+    @Nested
     @DisplayName("drop")
     class Drop {
 
@@ -341,6 +354,24 @@ class NonEmptyFiniteIterableTest {
     }
 
     @Nested
+    @DisplayName("foldRight")
+    class FoldRight {
+
+        @Test
+        void throwsOnNullOperator() {
+            FiniteIterable<Integer> ints = nonEmptyFiniteIterable(1, asList(2, 3));
+            assertThrows(NullPointerException.class, () -> ints.foldRight(null, lazy(0)));
+        }
+
+        @Test
+        void onSize5() {
+            FiniteIterable<String> items = nonEmptyFiniteIterable("1", asList("2", "3", "4", "5"));
+            assertEquals("6,5,4,3,2,1", items.foldRight((x, acc) -> acc.fmap(s -> s + "," + x), lazy("6")).value());
+        }
+
+    }
+
+    @Nested
     @DisplayName("inits")
     class Inits {
 
@@ -447,6 +478,43 @@ class NonEmptyFiniteIterableTest {
         void testCase2() {
             assertThat(nonEmptyFiniteIterable("foo", asList("bar", "baz")).prependAll("*"),
                     contains("*", "foo", "*", "bar", "*", "baz"));
+        }
+
+    }
+
+    @Nested
+    @DisplayName("reduceLeft")
+    class ReduceLeft {
+
+        @Test
+        void throwsOnNullOperator() {
+            NonEmptyFiniteIterable<Integer> ints = nonEmptyFiniteIterable(1, asList(2, 3));
+            assertThrows(NullPointerException.class, () -> ints.reduceLeft(null));
+        }
+
+        @Test
+        void onSize5() {
+            NonEmptyFiniteIterable<Integer> ints = nonEmptyFiniteIterable(1, asList(2, 3, 4, 5));
+            assertEquals(15, ints.reduceLeft(Integer::sum));
+        }
+
+    }
+
+    @Nested
+    @DisplayName("reduceRight")
+    class ReduceRight {
+
+        @Test
+        void throwsOnNullOperator() {
+            NonEmptyFiniteIterable<Integer> ints = nonEmptyFiniteIterable(1, asList(2, 3));
+            //noinspection ConstantConditions
+            assertThrows(NullPointerException.class, () -> ints.reduceRight(null));
+        }
+
+        @Test
+        void onSize5() {
+            NonEmptyFiniteIterable<String> items = nonEmptyFiniteIterable("1", asList("2", "3", "4", "5"));
+            assertEquals("5,4,3,2,1", items.reduceRight((x, acc) -> acc + "," + x));
         }
 
     }